% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_bia.R
\name{browse_bia}
\alias{browse_bia}
\title{Display an interactive map of BIA feature locations and associated photos}
\usage{
browse_bia(
  BIA_PATH,
  impact_values = 1:3,
  feature_layer = "BIAver2",
  feature_key = "GlobalID",
  attachment_layer = "BIAver2__ATTACH",
  attachment_key = "REL_GLOBALID",
  debug_info = FALSE
)
}
\arguments{
\item{impact_values}{(integer) An optional vector of one or more values. If
provided, BIA records will be filtered to those whose
\code{ImpactAssessment} value is in this set. Default is to filter on
values 1, 2 and 3. Set to \code{NULL} or an empty vector for no filtering
of records.}

\item{feature_layer}{(character) Name of the point feature layer in the
GeoDatabase for BIA locations. Default value is \code{'BIAver2'}.}

\item{feature_key}{(character) Name of the feature layer key field. Default
value is \code{'globalid'}.}

\item{attachment_layer}{(character) Name of the GeoDatabase table of
attachments. Default value is \code{'BIAver2__ATTACH'}.}

\item{attachment_key}{(character) Name of the attachment layer key field
linking records to features. Default value is \code{'rel_globalid'}.}

\item{debug_info}{(logical) If \code{TRUE}, print SQL query commands and
other internal information for debugging purposes. Default is \code{FALSE}
for no debug information.}

\item{BIA_path}{(character) Path to the ESRI geodatabase of BIA features and
associated attachments.}
}
\description{
This function launches a Shiny app with a map of BIA feature locations.
Clicking on a feature will query attachments table in the BIA database to
retrieve associated photographs and display them.
}
\examples{
\dontrun{
browse_bia(BIA_PATH = "data/BIA_SQL2008.gdb", impact_values = 1:3)
}

}
