% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_functions.R
\name{get_bia_photos}
\alias{get_bia_photos}
\title{Extract attached photos for selected BIA features}
\usage{
get_bia_photos(
  query_bounds,
  BIA_PATH,
  dest_dir,
  feature_layer = "BIAver2",
  feature_key = "globalid",
  attachment_layer = "BIAver2__ATTACH",
  attachment_key = "rel_globalid"
)
}
\arguments{
\item{dest_dir}{(character) Directory in which to write the JPG files for
photos. If the provided directory path does not exist it will be created.}

\item{feature_layer}{(character) Name of the point feature layer in the
GeoDatabase for BIA locations. Default value is \code{'BIAver2'}.}

\item{feature_key}{(character) Name of the feature layer key field. Default
value is \code{'globalid'}.}

\item{attachment_layer}{(character) Name of the GeoDatabase table of
attachments. Default value is \code{'BIAver2__ATTACH'}.}

\item{attachment_key}{(character) Name of the attachment layer key field
linking records to features. Default value is \code{'rel_globalid'}.}

\item{bounds}{An \code{sf} spatial data frame containing one or more polygon
features that define the bounds to be queried.}

\item{BIA_path}{(character) Path to the ESRI geodatabase of BIA features and
associated attachments.}
}
\value{
An sf spatial data frame of the BIA point features within the query
  bounds with all fields from the BIA table plus the column
  \code{'num_photos'} with number of photos associated with each feature.
}
\description{
Extract attached photos for selected BIA features
}
\examples{
\dontrun{
# Read in polygon(s) defining the query bounds
dat_bounds <- sf::st_read("conjola_boundary.shp")

# Get the photos for BIA features within this query area.
dat_query_features <- get_bia_photos(dat_bounds, BIA_PATH = "data/BIA_SQL2008.gdb", dest_dir = "conjola")
}

}
